Re-organise crate
